refactor hook folders
