--- code/system.ui/react.util/src/common.t.ts
@@ -1,2 +1,0 @@
-import type * as t from './common/types.mjs';
-export type { t };
