+++ code/system.ui/react.util/src/index.mts
@@ -1,3 +1,3 @@
-export { FC } from './ui/FC/index.mjs';
+export { FC } from './ui/FC/index';
 export { useClickInside, useClickOutside } from './use/use.Click';
 export { useMouse } from './use/use.Mouse';
