+++ code/system.ui/react.util/src/use/use.Mouse/types.mts
@@ -1,1 +1,8 @@
-export {};
+type M = React.MouseEventHandler;
+
+export type UseMouseProps = {
+  onDown?: M;
+  onUp?: M;
+  onEnter?: M;
+  onLeave?: M;
+};
